refactor(#176): split models/network.py - scanner config moves to models/scanner.py
Separates two concerns that were drifting in network.py:

* models/network.py keeps the asset-layer network properties -
  AsnRecord today, future RdapRecord (#180) / IpAsset / Netblock /
  RIROrganization as the amass-integration story lands. These describe
  the ASSET (what is this host / IP / netblock?).
* models/scanner.py (new) carries the CLI-scanner config + result
  shapes - NmapMode / NmapBanner / NmapScripts / NmapService /
  NmapHostResult / NmapScanResult / HttpxMode. These describe the
  SCAN (what did this nmap / httpx invocation ask and find?).

Same file naming logic the existing per-domain split already uses
(finding.py vs report.py, etc) - one concern per file, the next
contributor knows where to add things.

Import sweep across 9 files. The asn.py modules stay pointed at
models/network.py (AsnRecord still lives there); the nmap package
(flags / parser / scanner) and httpx.py + their tests retarget at
models/scanner.py. models/__init__.py re-exports from both with
updated table-of-contents.

Test surface: 1759 still passing (no behaviour change - this is a
pure restructure). Ruff + mypy clean.
