+++ models/__init__.py
@@ -20,8 +20,9 @@ across every consumer.
 | ``models.insight`` | ``HostAnnotation``, ``InsightValidationIssue``, |
 |                    | ``InsightValidationReport``, ``ReconFinalisationError`` |
 | ``models.metrics`` | ``RunMetrics`` |
-| ``models.network`` | ``AsnRecord``, ``NmapMode``, ``NmapBanner``, ``NmapScripts``, |
-|                    | ``NmapService``, ``NmapHostResult``, ``NmapScanResult`` |
+| ``models.network`` | ``AsnRecord`` |
+| ``models.scanner`` | ``NmapMode``, ``NmapBanner``, ``NmapScripts``, ``NmapService``, |
+|                    | ``NmapHostResult``, ``NmapScanResult``, ``HttpxMode`` |
 | ``models.h1`` | HackerOne shapes incl. ``ProgrammeReportSummary`` |
 | ``models.attack`` | ``AttackPlan``, ``AttackPlanItem``, |
 |                   | ``AttackPlanValidationIssue``, ``AttackPlanValidationReport``, |
@@ -67,8 +68,11 @@ from models.insight import (
     ReconFinalisationError,
 )
 from models.metrics import RunMetrics
-from models.network import (
-    AsnRecord,
+from models.network import AsnRecord
+from models.owasp import OWASPEntry
+from models.primitives import FQDN, HttpUrl, IPAddress, Severity
+from models.report import AuthoredDraft
+from models.scanner import (
     HttpxMode,
     NmapBanner,
     NmapHostResult,
@@ -77,9 +81,6 @@ from models.network import (
     NmapScripts,
     NmapService,
 )
-from models.owasp import OWASPEntry
-from models.primitives import FQDN, HttpUrl, IPAddress, Severity
-from models.report import AuthoredDraft
 from models.technology import Technology, TechnologyCategory
 from models.triage import AuthoredAssessment, SeverityDecision
 from models.workspace import RunFile, RunFileContent
